+++ src/controller/src/controller/mapper/account_mapper.cpp
@@ -1,10 +1,10 @@
-#include "drafts/account_mapper.hpp"
+#include "account_mapper.hpp"
 
 #include "config/id_types.hpp"
-#include "drafts/account_draft.hpp"
-#include "finance/account.hpp"
+#include "drafts/account/account_draft.hpp"
+#include "logic/finance/account.hpp"
 
-namespace drafts
+namespace controller
 {
 
     /**
@@ -19,15 +19,34 @@ namespace drafts
      * as the finance::Account but may be structured differently to better suit
      * the needs of the UI.
      */
-    AccountDraft AccountMapper::toDraft(const finance::Account& account)
+    drafts::AccountDraft AccountMapper::toDraft(const finance::Account& account)
     {
-        return AccountDraft{
+        return drafts::AccountDraft{
             account.getId(),
             account.getName(),
             account.getKind(),
             account.getCurrency(),
             account.getStatus()
         };
+    }
+
+    /**
+     * @brief Convert a vector of finance::Account to a vector of AccountDraft
+     *
+     * @param accounts The vector of finance::Account to convert
+     * @return std::vector<drafts::AccountDraft> The vector of corresponding
+     * AccountDraft
+     */
+    std::vector<drafts::AccountDraft> AccountMapper::toDrafts(
+        const std::vector<finance::Account>& accounts
+    )
+    {
+        std::vector<drafts::AccountDraft> drafts;
+        drafts.reserve(accounts.size());
+        for (const auto& account : accounts)
+            drafts.push_back(toDraft(account));
+
+        return drafts;
     }
 
     /**
@@ -42,9 +61,10 @@ namespace drafts
      * domain model representation of an account, and contains all the details
      * of the account as it exists in the business logic layer.
      */
-    finance::Account AccountMapper::toAccount(const AccountDraft& draft)
+    finance::Account AccountMapper::fromDraft(const drafts::AccountDraft& draft)
     {
         auto account = finance::Account{
+            AccountId::invalid(),
             draft.status.value_or(AccountStatus::Active),
             draft.name,
             draft.currency,
@@ -54,4 +74,4 @@ namespace drafts
         return account;
     }
 
-}   // namespace drafts
+}   // namespace controller