+++ src/controller/src/controller/side_bar/account_controller.cpp
@@ -5,8 +5,10 @@
 
 #include "commands/account/create_account_command.hpp"
 #include "commands/undo_stack.hpp"
+#include "controller/mapper/account_mapper.hpp"
+#include "drafts/account/account_draft.hpp"
 #include "logging/log_macros.hpp"
-#include "store/account/account_store.hpp"
+#include "store/i_account_store.hpp"
 #include "ui/account/create_account_dlg.hpp"
 #include "ui/side_bar/account_category.hpp"
 #include "ui/side_bar/account_item.hpp"
@@ -43,10 +45,10 @@ namespace controller
      *
      */
     AccountSideBarController::AccountSideBarController(
-        cmd::UndoStack&      undoStack,
-        store::AccountStore& accountStore,
-        AccountController&   accountController,
-        QMainWindow*         mainWindow
+        cmd::UndoStack&                              undoStack,
+        const std::shared_ptr<store::IAccountStore>& accountStore,
+        AccountController&                           accountController,
+        QMainWindow*                                 mainWindow
     )
         : SideBarCategoryController(new ui::AccountCategory(), mainWindow),
           _undoStack(undoStack),
@@ -69,13 +71,13 @@ namespace controller
             return;
 
         category->clearAccounts();
-        const auto accounts = _accountStore.getAllAccounts();
+        const auto accounts = _accountStore->getAllAccounts();
 
         for (const auto& account : accounts)
         {
             category->addAccount(
-                account.id,
-                QString::fromStdString(account.name)
+                account.getId(),
+                QString::fromStdString(account.getName())
             );
         }
     }
@@ -149,7 +151,7 @@ namespace controller
 
         auto result = cmd::Commands::makeAndDo<cmd::CreateAccountCommand>(
             _accountStore,
-            account
+            AccountMapper::fromDraft(account)
         );
 
         if (!result)