+++ src/store/include/store/base/base_store.hpp
@@ -225,6 +225,8 @@ namespace store
 
         void _logCache(const std::string& category, LogLevel level);
 
+        [[nodiscard]] const IdMap& _getIdRemap() const;
+
        private:
         static bool _evalDeletionPolicy(
             const Entry&   entry,