+++ src/ui/src/ui/account/create_account_dlg.cpp
@@ -8,7 +8,7 @@
 
 #include "config/finance.hpp"
 #include "config/id_types.hpp"
-#include "drafts/account_draft.hpp"
+#include "drafts/account/account_draft.hpp"
 #include "ui/validators/name_line_edit.hpp"
 #include "ui/validators/validators.hpp"
 #include "utils/qt_helpers.hpp"